add Rest assured test + add library for spring
