+++ pom.xml
@@ -206,6 +206,13 @@
       		<version>3.0.5</version>
       <scope>test</scope>
 		</dependency>
+		
+		<dependency>
+      		<groupId>io.rest-assured</groupId>
+      		<artifactId>spring-mock-mvc</artifactId>
+     		<version>3.0.5</version>
+      		<scope>test</scope>
+		</dependency>
     </dependencies>
 
     <build>
